requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
cairocffi==0.8.0
CairoSVG==2.1.3
celery==3.1.18
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==1.11.10
django-paypal==0.4.1
django-tinymce==2.7.0
django-wysiwyg-redactor==0.5.1
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
kombu==3.0.37
pdfrw==0.4
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycparser==2.18
Pyphen==0.9.4
pytz==2018.3
requests==2.18.4
six==1.11.0
tinycss2==0.6.1
urllib3==1.22
WeasyPrint==0.42.2
webencodings==0.5.1
uwsgi==2.0.17
mysqlclient==1.3.12
pycrypto==2.6.1

